Rips out Postgres - but leaves migration files for now
- this does not yet remove the models
- uses sqlite3 driver
- changes the config based on sqlite3
- tests pass but no depth in investigating the models

The sqlite3 database is going to be used for storing user data and
tokens only - we are not going to store the models in a database at this
point.

If we want to start saving datamaps, datamaplines, returns and
returnlines, etc, we will need the models but as of this commit the
models are still in the code.